refactor: migrate webhook payload handling to dedicated module
